refactor(core): split benchmark domain module into per-concern files
Convert domain/benchmark.rs into a domain/benchmark/ module directory,
mirroring the existing domain/agent/ split (config.rs/events.rs/messages.rs).
Pure move: run.rs (BenchmarkRun/Status/Type), summary.rs
(ModelBenchmarkSummary), compare.rs (CompareConfig/ModelCompareResult),
perf.rs (PerfConfig/ModelPerfResult), events.rs (BenchmarkEvent/
BenchmarkModelResult). No behavior or public API changes. Adds the
required module README with module-docs/module-table markers.

Prepares the ground for the upcoming tune/ submodule (benchmark
parameter-sweep + agentic tool-calling eval) without mixing new
functionality into this refactor.
